requirements.txt
piexif==1.1.3
python-dateutil==2.8.1
selenium==3.141.0
wget==3.2
